+++ komikaan.Data/komikaan.Data.csproj
@@ -13,4 +13,8 @@
     </PackageReference>
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\GTFS\src\GTFS\GTFS.csproj" />
+  </ItemGroup>
+
 </Project>
